power_domain: Add optional configuration for composite state mask
Extending the composite state is required to support
device implementation defined composite state.

This patch adds a new configuration option to provide composite state
mask table for each power domain element. The configuration is optional.
If the table is not provided then composite state is not supported.
The core power domain is a special case. If the table is not provided
for the core the default table is used.

Change-Id: I7b20c02414f8d11d3606a7829470af15d18af6ff
Signed-off-by: Tarek El-Sherbiny <tarek.el-sherbiny@arm.com>